move country selector to _app
