+++ pages/holidays.js
@@ -1,7 +1,6 @@
 import { useRouter } from 'next/router'
 import { useState, useEffect } from 'react'
 import getConfig from 'next/config'
-import CountrySelector from '../components/CountrySelector'
 
 const Holidays = (props) => {
 
@@ -32,7 +31,6 @@ const Holidays = (props) => {
 
   return  (
     <div>
-      <CountrySelector />
       <ul>
         {holidays.map(holiday => <li key={holiday.uuid} value={holiday.uuid}>{holiday.date} - {holiday.name}</li>)}
       </ul>
